requirements.txt
streamlit
openai
requests
python-dotenv
pandas
numpy
flask
transformers
